Ignore old student enrollments in paces api
- Only allow operations for the student's most recent enrollment in
the paces API, the show, create, update, and delete endpoints will
return 404 when attempting to do operations with enrollments other
than the student's most recent

- Course_pacing#index sets only the most recent enrollment of each
student to ENV[ENROLLMENTS]

closes LS-3514
flag=course_paces_for_students

Test plan:
- In a course with course pacing
- Enroll a student in multiple sections

Using the most recent student_enrollment

- Make the following requests, and expect the endpoints to work as usual:
GET
api/v1/courses/:course_id/student_enrollments/:student_enrollment_id/pace
POST
api/v1/courses/:course_id/student_enrollments/:student_enrollment_id/paces
PATCH
api/v1/courses/:course_id/student_enrollments/:student_enrollment_id/pace
DELETE
api/v1/courses/:course_id/student_enrollments/:student_enrollment_id/pace

Using any other student_enrollment of the same user in the course

- Repeat the requests and expect to get a 404 for all of them

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/308819
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jonathan Guardado <jonathan.guardado@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>

+++ app/controllers/course_pacing/paces_api_controller.rb
@@ -26,6 +26,8 @@ class CoursePacing::PacesApiController < ApplicationController
 
   def show
     pace = pacing_service.pace_for(context, should_duplicate: true)
+    return not_found if pace.nil?
+
     render json: {
       pace: pacing_presenter.new(pace).as_json,
       progress: pace.id ? progress_json(pacing_service.progress(pace, publish: false), @current_user, session) : nil
@@ -34,6 +36,8 @@ class CoursePacing::PacesApiController < ApplicationController
 
   def create
     pace = pacing_service.create_in_context(context)
+    return not_found if pace.nil?
+
     render json: {
       pace: pacing_presenter.new(pace).as_json,
       progress: progress_json(pacing_service.progress(pace), @current_user, session)
@@ -42,6 +46,8 @@ class CoursePacing::PacesApiController < ApplicationController
 
   def update
     pace = pacing_service.pace_in_context(context)
+    return not_found if pace.nil?
+
     if pacing_service.update_pace(pace, update_params)
       render json: {
         pace: pacing_presenter.new(pace).as_json,
@@ -53,6 +59,10 @@ class CoursePacing::PacesApiController < ApplicationController
   end
 
   def delete
+    # make sure the pace exists and it is valid
+    pace = pacing_service.pace_in_context(context)
+    return not_found if pace.nil?
+
     pacing_service.delete_in_context(context)
     head :no_content
   end
